requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
python-multipart
python-dotenv
httpx
databases
asyncpg
Pillow
python-barcode[images]